+++ test/init/discovery_test.clj
@@ -1,6 +1,9 @@
 (ns init.discovery-test
-  (:require [clojure.test :refer [deftest is]]
+  (:require [clojure.string :as str]
+            [clojure.test :refer [deftest is]]
             [init.discovery :as discovery]))
+
+(def this-ns (ns-name *ns*))
 
 (deftest ns-prefix-pred-test
   (let [pred (#'discovery/ns-prefix-pred 'init.discovery-test)]
@@ -11,11 +14,12 @@
     (is (not (pred 'init.discovery-test-sibling)))))
 
 (deftest classpath-namespaces-test
-  (let [namespaces (discovery/classpath-namespaces)]
-    (is (contains? namespaces (ns-name *ns*)))
-    (is (contains? namespaces 'init.core)))
   (let [namespaces (discovery/classpath-namespaces ['init])]
-    (is (contains? namespaces (ns-name *ns*)))
+    (is (contains? namespaces this-ns))
     (is (contains? namespaces 'init.core)))
   (let [namespaces (discovery/classpath-namespaces ['init.core 'init.foobar])]
-    (is (empty? namespaces))))
+    (is (empty? namespaces)))
+  (let [namespaces (discovery/classpath-namespaces ['init] :include? #(str/ends-with? % "core"))]
+    (is (= ['init.core] (vec namespaces))))
+  (let [namespaces (discovery/classpath-namespaces ['init] :exclude? #(str/ends-with? % "-test"))]
+    (is (not (contains? namespaces this-ns)))))